Core migration: IEnumerableExtensions.IsNullOrEmpty
